+++ py/dml/expr.py
@@ -7,6 +7,7 @@ import dml.globals
 from . import logging
 from .logging import ICE
 from .messages import *
+from . import errors as E
 from . import output
 from . import types as tp
 from .slotsmeta import SlotsMeta, auto_init
@@ -202,7 +203,7 @@ class NonValue(Expression):
         raise self.exc()
     def exc(self):
         '''Exception to raise when expression appears in an incorrect context'''
-        return ENVAL(self.site, self)
+        return E.NVAL(self.site, self)
 
 class NonValueArrayRef(NonValue):
     '''Reference to an array node before it's indexed. Indexing is the
@@ -215,7 +216,7 @@ class NonValueArrayRef(NonValue):
     def local_dimsizes(self): pass
 
     def exc(self):
-        return EARRAY(self.site, self)
+        return E.ARRAY(self.site, self)
 
 class Lit(Expression):
     "A literal C expression"
@@ -265,7 +266,7 @@ mkNullConstant = NullConstant
 def typecheck_inargs(site, args, inp, kind="function", known_arglen=None):
     arglen = len(args) if known_arglen is None else known_arglen
     if arglen != len(inp):
-        raise EARG(site, kind)
+        raise E.ARG(site, kind)
 
     for (i, (arg, p)) in enumerate(zip(args, inp)):
         if kind == 'method':
@@ -283,9 +284,9 @@ def typecheck_inargs(site, args, inp, kind="function", known_arglen=None):
         (ok, trunc, constviol) = rtype.canstore(argtype)
         if ok:
             if constviol:
-                raise ECONSTP(site, logref, kind + " call")
+                raise E.CONSTP(site, logref, kind + " call")
         else:
-            raise EPTYPE(site, arg, rtype, logref, kind)
+            raise E.PTYPE(site, arg, rtype, logref, kind)
 
 # Typecheck a DML method application, where the arguments are given as a list
 # where each element is either an AST of an initializer, or an initializer
@@ -296,7 +297,7 @@ def typecheck_inarg_inits(site, inits, inp, location, scope,
                           allow_undefined_args=False,
                           on_ptr_to_stack=None):
     if (not variadic and len(inits) != len(inp)) or len(inits) < len(inp):
-        raise EARG(site, kind)
+        raise E.ARG(site, kind)
 
     from .expr_util import coerce_if_eint
     from .codegen import eval_initializer, codegen_expression, \
@@ -319,19 +320,19 @@ def typecheck_inarg_inits(site, inits, inp, location, scope,
             else:
                 try:
                     arg = init.as_expr(ptype)
-                except EASTYPE as e:
+                except E.ASTYPE as e:
                     if e.site is init.site:
-                        raise EPTYPE(site, e.source, e.target_type, logref,
+                        raise E.PTYPE(site, e.source, e.target_type, logref,
                                      kind) from e
                     raise
                 # better error message
-                except EDISCONST as e:
+                except E.DISCONST as e:
                     if e.site is init.site:
-                        raise ECONSTP(site, logref, kind + " call") from e
+                        raise E.CONSTP(site, logref, kind + " call") from e
                     raise
         elif ptype is None:
             if init.kind != 'initializer_scalar':
-                raise ESYNTAX(init.site, '{',
+                raise E.SYNTAX(init.site, '{',
                               'the argument for an untyped parameter must be '
                               + 'a simple expression')
             arg = codegen_expression_maybe_nonvalue(init.args[0], location,
@@ -354,22 +355,22 @@ def typecheck_inarg_inits(site, inits, inp, location, scope,
 
             if ok:
                 if constviol:
-                    raise ECONSTP(site, logref, kind + " call")
+                    raise E.CONSTP(site, logref, kind + " call")
             else:
-                raise EPTYPE(site, arg, rtype, logref, kind)
+                raise E.PTYPE(site, arg, rtype, logref, kind)
         else:
             try:
                 arg = eval_initializer(init.site, ptype, init, location,
                                        scope, False).as_expr(ptype)
-            except EASTYPE as e:
+            except E.ASTYPE as e:
                 if e.site is init.site:
-                    raise EPTYPE(site, e.source, e.target_type, logref,
+                    raise E.PTYPE(site, e.source, e.target_type, logref,
                                  kind) from e
                 raise
             # better error message
-            except EDISCONST as e:
+            except E.DISCONST as e:
                 if e.site is init.site:
-                    raise ECONSTP(site, logref, kind + " call") from e
+                    raise E.CONSTP(site, logref, kind + " call") from e
                 raise
         if (on_ptr_to_stack
             and isinstance(tp.safe_realtype_shallow(ptype), tp.Ptr)
@@ -380,7 +381,7 @@ def typecheck_inarg_inits(site, inits, inp, location, scope,
     if variadic and len(inits) > len(inp):
         for init in inits[len(inp):]:
             if init.kind != 'initializer_scalar':
-                raise ESYNTAX(init.site, '{',
+                raise E.SYNTAX(init.site, '{',
                               'variadic arguments must be simple expressions')
             args.append(coerce_if_eint(codegen_expression(init.args[0],
                                                           location, scope)))
@@ -407,7 +408,7 @@ def mkApplyInits(site, fun, inits, location, scope):
     funtype = fun.ctype()
 
     if not funtype:
-        raise EAPPLY(fun)
+        raise E.APPLY(fun)
 
     try:
         funtype = tp.realtype(funtype)
@@ -415,10 +416,10 @@ def mkApplyInits(site, fun, inits, location, scope):
             # Pointers to functions are the same as the functions
             funtype = tp.realtype(funtype.base)
     except tp.DMLUnknownType:
-        raise ETYPE(site, funtype)
+        raise E.TYPE(site, funtype)
 
     if not isinstance(funtype, tp.Function):
-        raise EAPPLY(fun)
+        raise E.APPLY(fun)
 
     args = typecheck_inarg_inits(
         site, inits,
@@ -432,7 +433,7 @@ def mkApply(site, fun, args):
     funtype = fun.ctype()
 
     if not funtype:
-        raise EAPPLY(fun)
+        raise E.APPLY(fun)
 
     try:
         funtype = tp.realtype(funtype)
@@ -440,10 +441,10 @@ def mkApply(site, fun, args):
             # Pointers to functions are the same as the functions
             funtype = tp.realtype(funtype.base)
     except tp.DMLUnknownType:
-        raise ETYPE(site, funtype)
+        raise E.TYPE(site, funtype)
 
     if not isinstance(funtype, tp.Function):
-        raise EAPPLY(fun)
+        raise E.APPLY(fun)
 
     if funtype.varargs and len(args) > len(funtype.input_types):
         known_arglen = len(funtype.input_types)
@@ -482,4 +483,4 @@ class StaticIndex(NonValue):
     def __str__(self):
         return logging.dollar(self.site) + ("_" if self.var is None else self.var)
     def exc(self):
-        return EIDXVAR(self.site, str(self))
+        return E.IDXVAR(self.site, str(self))
