+++ py/dml/logging.py
@@ -26,6 +26,9 @@ __all__ = (
     'InEachSite',
 
     'dbg',
+
+    'truncate',
+    'binary_dump',
     )
 
 import sys
@@ -119,6 +122,8 @@ class LogMessage(object):
 
     outfile = sys.stderr
 
+    tag_prefix = ''
+
     def __init__(self, site, *msgargs):
         # The site is the place in the source that this log message
         # refers to.
@@ -156,8 +161,9 @@ class LogMessage(object):
             if os.getenv('DMLC_DEBUG'):
                 raise
 
-    def tag(self):
-        return self.__class__.__name__
+    @classmethod
+    def tag(cls):
+        return cls.tag_prefix + cls.__name__
 
     def preprocess(self):
         '''Call before log when reporting. Return True to actually log
@@ -214,6 +220,7 @@ class ICE(Exception, LogMessage):
 # This is a base class for warning messages
 #
 class DMLWarning(LogMessage):
+    tag_prefix = ''
     kind = "warning"
     next_warning_yields_error = False
 
@@ -239,6 +246,7 @@ class DMLWarning(LogMessage):
 # This is a base class for error messages
 #
 class DMLError(Exception, LogMessage):
+    tag_prefix = 'E'
     kind = "error"
 
     def __init__(self, site, *msgargs):
@@ -250,6 +258,7 @@ class DMLError(Exception, LogMessage):
         report_error()
 
 class PortingMessage(LogMessage):
+    tag_prefix = ''
     kind = 'porting'
     fmt = ''
 
@@ -487,3 +496,17 @@ def suppress_errors():
         store_errors.append(e)
     finally:
         store_errors = orig
+
+def truncate(s, maxlen):
+    "Make sure that s is not longer than maxlen"
+    if len(s) > maxlen:
+        return s[:maxlen-3] + '...'
+    return s
+
+def binary_dump(lh, rh):
+    """Produce a string to use in warning and error messages describing
+    operands to a binary operation"""
+    return ("LH: '%s' of type '%s'\n"
+            "RH: '%s' of type '%s'"
+            % (truncate(str(lh), 40), lh.ctype(),
+               truncate(str(rh), 40), rh.ctype()))
